Simplify IOPlugin.Data initialization
Remove the need to specify every component of Data. Note that after this
change, the IOPlugin.persisted interface may return None as falsive
value. Since there is no identity check (x is False), it is fine.

Also remove account_key and key from FullChainFile, it is not stored by
ChainFile.